+++ semcore/date-picker/src/DateRangePicker.jsx
@@ -106,6 +106,7 @@ class DateRangePickerRoot extends RangePickerAbstract {
       children: () => <InputTrigger.DateRange />,
       getI18nText,
       animationsDisabled,
+      unit: this.keyStep,
     };
   }
 }